Removed cleanDatabase-API

+++ lib/database.go
@@ -96,6 +96,10 @@ func prepareDatabase() {
 
 // Removes all check-results older than one month from the Database
 func CleanDatabase() {
+	if config.CheckLifetime <= 0 {
+		return
+	}
+
 	res, err := db.Exec("DELETE FROM checks WHERE time <= NOW() - INTERVAL ? DAY;", config.CheckLifetime)
 	if err != nil {
 		logging.MustGetLogger("").Fatal("Unable to cleanup Database: ", err)
@@ -103,7 +107,7 @@ func CleanDatabase() {
 	}
 
 	rowsAffected, _ := res.RowsAffected()
-	logging.MustGetLogger("").Info("Cleaned " + strconv.FormatInt(rowsAffected, 10) + " rows from the Database.")
+	logging.MustGetLogger("").Info("Cleaned " + strconv.FormatInt(rowsAffected, 10) + " rows from the Database (older than " + strconv.Itoa(config.CheckLifetime) + " days).")
 }
 
 // Returns the current Database-object.
